+++ Backend/src/ConsertaPraMim.Application/DependencyInjection.cs
@@ -15,6 +15,7 @@ public static class DependencyInjection
         services.AddScoped<ConsertaPraMim.Application.Interfaces.IServiceAppointmentService, ConsertaPraMim.Application.Services.ServiceAppointmentService>();
         services.AddScoped<ConsertaPraMim.Application.Interfaces.IServiceAppointmentNoShowRiskService, ConsertaPraMim.Application.Services.ServiceAppointmentNoShowRiskService>();
         services.AddScoped<ConsertaPraMim.Application.Interfaces.IAppointmentReminderService, ConsertaPraMim.Application.Services.AppointmentReminderService>();
+        services.AddScoped<ConsertaPraMim.Application.Interfaces.IGoogleCalendarSyncOperationsService, ConsertaPraMim.Application.Services.GoogleCalendarSyncOperationsService>();
         services.AddScoped<ConsertaPraMim.Application.Interfaces.IProposalService, ConsertaPraMim.Application.Services.ProposalService>();
         services.AddScoped<ConsertaPraMim.Application.Services.ReviewService>();
         services.AddScoped<ConsertaPraMim.Application.Interfaces.IReviewService>(